Format statistics and report output, remove un-necessary zeros

+++ stats/query-results-stats.go
@@ -1,30 +1,30 @@
 package stats
 
 type Statistical interface {
-	GetStatisticsMap() map[string]int
+	GetStatisticsMap() map[string]float64
 }
 
-func CalculateStats(statisticalRecords []Statistical) (map[string]int, map[string]float64, map[string]int, map[string]int) {
+func CalculateStats(statisticalRecords []Statistical) (map[string]float64, map[string]float64, map[string]float64, map[string]float64) {
 
-	var sum map[string]int = make(map[string]int)
+	var sum map[string]float64 = make(map[string]float64)
 	var avg map[string]float64 = make(map[string]float64)
-	var min map[string]int = make(map[string]int)
-	var max map[string]int = make(map[string]int)
+	var min map[string]float64 = make(map[string]float64)
+	var max map[string]float64 = make(map[string]float64)
 
 	var numberOfRecords int = len(statisticalRecords)
 
 	for index, record := range statisticalRecords {
 
 		for key, value := range record.GetStatisticsMap() {
-			sum[key] += value
+			sum[key] += float64(value)
 			avg[key] += float64(value)
 
-			if min[key] > value {
-				min[key] = value
+			if min[key] > float64(value) {
+				min[key] = float64(value)
 			}
 
-			if max[key] < value {
-				max[key] = value
+			if max[key] < float64(value) {
+				max[key] = float64(value)
 			}
 
 			// If this is the last record, calculate the average
@@ -33,10 +33,6 @@ func CalculateStats(statisticalRecords []Statistical) (map[string]int, map[strin
 			}
 		}
 	}
-
-	//for key, value := range sum {
-	//	fmt.Printf("Sum %s = %d\n", key, value)
-	//}
 
 	return sum, avg, min, max
 }
